Fix the Flow.js 0.88 changes
