+++ definitions/npm/redux-saga_v0.16.x/flow_v0.56.0-/test_redux-saga_0.16.x.js
@@ -471,7 +471,7 @@ function callTest() {
   (c8.CALL.context: null);
 
   // $ExpectError
-  (c1.CALL.context: Object);
+  (c1.CALL.context: {});
 }
 
 function callNormalFunctionTest() {
@@ -540,7 +540,7 @@ function callNormalFunctionTest() {
   (c8.CALL.context: null);
 
   // $ExpectError
-  (c1.CALL.context: Object);
+  (c1.CALL.context: {});
 }
 
 function callSagaFunctionTest() {
@@ -609,7 +609,7 @@ function callSagaFunctionTest() {
   (c8.CALL.context: null);
 
   // $ExpectError
-  (c1.CALL.context: Object);
+  (c1.CALL.context: {});
 }
 
 function contextCallTest() {
