+++ definitions/npm/redux-saga_v1.x.x/flow_v0.76.0-/test_redux-saga_1.x.x_call_async.js
@@ -33,7 +33,7 @@ describe("call() effect with async functions and without context", () => {
 
     it("must raises an error when lead context to Object", () => {
       // $ExpectError
-      (c1.payload.context: Object);
+      (c1.payload.context: {});
     });
   });
