+++ definitions/npm/redux-saga_v1.x.x/flow_v0.76.0-/test_redux-saga_1.x.x_call_normal.js
@@ -1,75 +1,74 @@
 // @flow
 import { describe, it } from "flow-typed-test";
-import type { Saga } from "redux-saga";
 import { call } from "redux-saga/effects";
 
 describe("call() effect with normal function and without context", () => {
-  function* s0(): Saga<number> {
+  function nfn0(): number {
     return 1;
   }
-  function* s1(a: string): Saga<number> {
+  function nfn1(a: string): number {
     return 1;
   }
-  function* s2(a: string, b: number): Saga<number> {
+  function nfn2(a: string, b: number): number {
     return 1;
   }
-  function* s3(a: string, b: number, c: string): Saga<number> {
+  function nfn3(a: string, b: number, c: boolean): number {
     return 1;
   }
-  function* s4(a: string, b: number, c: string, d: number): Saga<number> {
+  function nfn4(a: string, b: number, c: boolean, d: string): number {
     return 1;
   }
-  function* s5(a: string, b: number, c: string, d: number, e: string): Saga<number> {
+  function nfn5(a: string, b: number, c: boolean, d: string, e: number): number {
     return 1;
   }
-  function* s6(a: string, b: number, c: string, d: number, e: string, f: number): Saga<number> {
+  function nfn6(a: string, b: number, c: boolean, d: string, e: number, f: boolean): number {
     return 1;
   }
-  function* s7(
+  function nfn7(
     a: string,
     b: number,
-    c: string,
-    d: number,
-    e: string,
-    f: number,
+    c: boolean,
+    d: string,
+    e: number,
+    f: boolean,
     g: string
-  ): Saga<number> {
+  ): number {
     return 1;
   }
-  function* s8(
+  function nfn8(
     a: string,
     b: number,
-    c: string,
-    d: number,
-    e: string,
-    f: number,
+    c: boolean,
+    d: string,
+    e: number,
+    f: boolean,
     g: string,
     h: number
-  ): Saga<number> {
+  ): number {
     return 1;
   }
 
-  const c0 = call(s0);
-  const c1 = call(s1, "1");
-  const c2 = call(s2, "1", 2);
-  const c3 = call(s3, "1", 2, "3");
-  const c4 = call(s4, "1", 2, "3", 4);
-  const c5 = call(s5, "1", 2, "3", 4, "5");
-  const c6 = call(s6, "1", 2, "3", 4, "5", 6);
-  const c7 = call(s7, "1", 2, "3", 4, "5", 6, "7");
-  const c8 = call(s8, "1", 2, "3", 4, "5", 6, "7", 8);
+  const c0 = call(nfn0);
+  const c1 = call(nfn1, "1");
+  const c2 = call(nfn2, "1", 2);
+  const c3 = call(nfn3, "1", 2, true);
+  const c4 = call(nfn4, "1", 2, true, "4");
+  const c5 = call(nfn5, "1", 2, true, "4", 5);
+  const c6 = call(nfn6, "1", 2, true, "4", 5, false);
+  const c7 = call(nfn7, "1", 2, true, "4", 5, false, "7");
+  const c8 = call(nfn8, "1", 2, true, "4", 5, false, "7", 8);
 
   describe("arguments tests", () => {
     it("must passes when used properly", () => {
       (c0.payload.args: []);
       (c1.payload.args: [string]);
       (c2.payload.args: [string, number]);
-      (c3.payload.args: [string, number, string]);
-      (c4.payload.args: [string, number, string, number]);
-      (c5.payload.args: [string, number, string, number, string]);
-      (c6.payload.args: [string, number, string, number, string, number]);
-      (c7.payload.args: [string, number, string, number, string, number, string]);
-      (c8.payload.args: [string, number, string, number, string, number, string, number]);
+      (c3.payload.args: [string, number, boolean]);
+      (c4.payload.args: [string, number, boolean, string]);
+      (c5.payload.args: [string, number, boolean, string, number]);
+      (c6.payload.args: [string, number, boolean, string, number, boolean]);
+      (c7.payload.args: [string, number, boolean, string, number, boolean, string]);
+      (c8.payload.args: [string, number, boolean, string, number, boolean, string, number]);
     });
 
     it("must raises an error when passed number but need string", () => {
@@ -79,47 +78,47 @@ describe("call() effect with normal function and without context", () => {
 
     it("must raises an error when passed too few arguments", () => {
       // $ExpectError: Too few arguments
-      call(s6, "1", 2, "3", 4);
+      call(nfn6, "1", 2, true, "4");
     });
 
     it("must raises an error when passed wrong argument types", () => {
       // $ExpectError: Wrong argument types
-      call(s1, 1);
+      call(nfn1, 1);
     });
   });
 
   describe("function test", () => {
     it("must passes when used properly", () => {
-      (c1.payload.fn: typeof s1);
-      (c2.payload.fn: typeof s2);
-      (c3.payload.fn: typeof s3);
-      (c4.payload.fn: typeof s4);
-      (c5.payload.fn: typeof s5);
-      (c6.payload.fn: typeof s6);
-      (c7.payload.fn: typeof s7);
-      (c8.payload.fn: typeof s8);
+      (c1.payload.fn: typeof nfn1);
+      (c2.payload.fn: typeof nfn2);
+      (c3.payload.fn: typeof nfn3);
+      (c4.payload.fn: typeof nfn4);
+      (c5.payload.fn: typeof nfn5);
+      (c6.payload.fn: typeof nfn6);
+      (c7.payload.fn: typeof nfn7);
+      (c8.payload.fn: typeof nfn8);
     });
 
     it("should actually fail, but apparently more parameter are fine", () => {
-      (c1.payload.fn: typeof s6);
+      (c1.payload.fn: typeof nfn6);
     });
 
     it("must raises an error when Function return not string", () => {
-      // $ExpectError: fn returns a Saga<number> not Saga<string>
-      (c1.payload.fn: (a: string) => Saga<string>);
+      // $ExpectError: fn returns a number not string
+      (c1.payload.fn: (a: boolean) => string);
     });
 
     it(`must raises an error when "a" argument isn't string`, () => {
       // $ExpectError: 'a' is actually of type string
-      (c1.payload.fn: (a: boolean) => Saga<number>);
+      (c1.payload.fn: (a: boolean) => number);
 
       // $ExpectError: 'a' is actually of type string
-      (c4.payload.fn: (a: number, b: number) => Saga<number>);
+      (c4.payload.fn: (a: number, b: number) => number);
     });
 
     it("must raises an error when less parameter are noticed", () => {
       // $ExpectError: Less parameter are noticed
-      (c6.payload.fn: typeof s1);
+      (c6.payload.fn: typeof nfn1);
     });
   });
 
@@ -137,7 +136,7 @@ describe("call() effect with normal function and without context", () => {
 
     it("must raises an error when lead context to Object", () => {
       // $ExpectError
-      (c1.payload.context: Object);
+      (c1.payload.context: {});
     });
   });
 });
