+++ definitions/npm/redux-saga_v1.x.x/flow_v0.76.0-/test_redux-saga_1.x.x_spawn.js
@@ -157,7 +157,7 @@ describe("spawn effect", () => {
 
       it("must raises an error when lead context to Object", () => {
         // $ExpectError
-        (c1.payload.context: Object);
+        (c1.payload.context: {});
       });
     });
   });
